pyproject.toml
[tool.poetry]
name = "streaming"
version = "0.1.0"
description = ""
authors = ["Denis Shchtskyi <denis.shchutskyi@branderstudio.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
pydantic-settings = "^2.0.3"
uvicorn = "^0.23.2"
starlette-prometheus = "^0.9.0"
aiortc = "^1.5.0"
asgiref = "^3.7.2"
aiohttp = "^3.8.5"
uvloop = "^0.17.0"


[tool.poetry.group.lint.dependencies]
isort = "^5.12.0"
ruff = "^0.0.292"
mypy = "^1.5.1"
black = "^23.9.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 88
include_trailing_comma = true
multi_line_output = 3
force_grid_wrap = 0
combine_as_imports = true

[tool.coverage.run]
branch = true
source = ["checklists", "tests"]
omit = ["site-packages"]

[tool.mypy]
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_subclassing_any = false
mypy_path = "stubs"

[[tool.mypy.overrides]]
module = "*"
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*/migrations/*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "requests.*"
ignore_missing_imports = true

[[tool.mypy-requests.overrides]]
module = "requests.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytz.*"
ignore_missing_imports = true

[tool.ruff]
select = [
    "E",
    "F",
    "W",
    "C90",
]
unfixable = ["ALL"]
ignore = [
    "E203",
    "F401",
]
line-length = 88
exclude = [
    ".git",
    ".pytest_cache",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "venv",
    "tests",
    '__init__.py'
]

[tool.ruff.per-file-ignores]
"*__init__.py" = ["F401"]
"*/migrations/*" = ["E501"]

